+++ CHANGELOG.md
@@ -40,6 +40,7 @@ For more information about this file see also [Keep a Changelog](http://keepacha
 - Added small container to check if certain services are up, used as initi container for kubernetes
 - Documentation how to run ED using singularity
 - PEcAn.DB gains new function `get_postgres_envvars`, which tries to look up connection parameters from Postgres environment variables (if they are set) and return them as a list ready to be passed to `db.open`. It should be especially useful when writing tests that need to run on systems with many different database configurations (#2541).
+- New shiny application to show database synchronization status (shiny/dbsync)
 
 ### Removed
 - Database maintenance scripts `vacuum.bety.sh` and `reindex.bety.sh` have been moved to the [BeTY database repository](https://github.com/PecanProject/bety) (#2563).